feat: migrate sticker service to PostgreSQL; update database configuration and schema

+++ services/sticker/db/tableSticker.js
@@ -42,9 +42,9 @@ function init(db) {
       ${columns.searchVisible} INTEGER NOT NULL DEFAULT 1,
       ${columns.status} TEXT NOT NULL DEFAULT '${stickerStatus.ACTIVE}',
       ${columns.sortOrder} INTEGER NOT NULL DEFAULT 0,
-      ${columns.createdAt} INTEGER NOT NULL,
-      ${columns.updatedAt} INTEGER NOT NULL,
-      ${columns.deletedAt} INTEGER DEFAULT NULL,
+      ${columns.createdAt} BIGINT NOT NULL,
+      ${columns.updatedAt} BIGINT NOT NULL,
+      ${columns.deletedAt} BIGINT DEFAULT NULL,
       CONSTRAINT fk_sticker_pack
         FOREIGN KEY (${columns.packId}) REFERENCES tableStickerPack(id)
         ON UPDATE CASCADE ON DELETE RESTRICT
