+++ services/sticker/db/tableStickerCategory.js
@@ -30,9 +30,9 @@ function init(db) {
       ${columns.previewStickerId} TEXT DEFAULT NULL,
       ${columns.status} TEXT NOT NULL DEFAULT '${categoryStatus.ACTIVE}',
       ${columns.sortOrder} INTEGER NOT NULL DEFAULT 0,
-      ${columns.createdAt} INTEGER NOT NULL,
-      ${columns.updatedAt} INTEGER NOT NULL,
-      ${columns.deletedAt} INTEGER DEFAULT NULL
+      ${columns.createdAt} BIGINT NOT NULL,
+      ${columns.updatedAt} BIGINT NOT NULL,
+      ${columns.deletedAt} BIGINT DEFAULT NULL
     );
 
     CREATE INDEX IF NOT EXISTS idx_sticker_category_status
